* commit '74954d763cbf88a644893853e8331b35ff1e2d9d':
  Convert jint to jlong for native pointers in GLFunctionTable to support 64 bit.

+++ chromium/plat_support/graphics_utils.cpp
@@ -105,17 +105,17 @@ bool IsSkiaVersionCompatible(SkiaVersionFunction function) {
   return compatible;
 }
 
-jint GetDrawSWFunctionTable(JNIEnv* env, jclass) {
+jlong GetDrawSWFunctionTable(JNIEnv* env, jclass) {
   static const AwDrawSWFunctionTable function_table = {
       &GetPixels,
       &ReleasePixels,
       &CreatePicture,
       &IsSkiaVersionCompatible,
   };
-  return reinterpret_cast<jint>(&function_table);
+  return reinterpret_cast<intptr_t>(&function_table);
 }
 
-jint GetDrawGLFunctionTable(JNIEnv* env, jclass) {
+jlong GetDrawGLFunctionTable(JNIEnv* env, jclass) {
   static const AwDrawGLFunctionTable function_table = {
     &GraphicBufferImpl::Create,
     &GraphicBufferImpl::Release,
@@ -124,7 +124,7 @@ jint GetDrawGLFunctionTable(JNIEnv* env, jclass) {
     &GraphicBufferImpl::GetNativeBufferStatic,
     &GraphicBufferImpl::GetStrideStatic,
   };
-  return reinterpret_cast<jint>(&function_table);
+  return reinterpret_cast<intptr_t>(&function_table);
 }
 
 const char kClassName[] = "com/android/webview/chromium/GraphicsUtils";
